+++ NEWS.md
@@ -1,5 +1,11 @@
 # DataTools
 
+## Version 23.03.2
+
+## New Features
+- _Remote Models_ module: UI and server functions for selecting a remote (or local) model and
+obtaining the remote path on github (or the local path) to the file of the selected model
+
 ## Version 23.03.1
 
 ## Updates